requirements.txt
python-dotenv
flask
authlib
flask_sqlalchemy
langchain
langchain-community
langchain-core
langchain-groq
psycopg2-binary